Add frame for public API including algorithm docu
This commit first and foremost adds the `Allocator` type, which is the
only exposed type in this crate. The type implements `GlobalAlloc` to
make it usable as a global allocator. The API is as minimal as possible
and there is no runtime introspection of the allocation state.

Another important part of the commit is the crate and type documentation
showing the usage and describing, how the allocation algorithm works.