Add the "raw allocator" abstraction
This commit creates a new module `raw_allocator` with the `RawAllocator`
type. This allocator is a simpler version of a global allocator as it
does not need to handle alignment. Therefore the main `Allocator` uses
that new allocator to alloc/free memory slices. The main allocator also
ensures the proper alignment.