Implement raw buffer handling & complete allocator
This is a rather huge commit, which implements the missing parts of the
allocator. The new part is the `buffer` module, which does all the
unsafe memory reading/writing/handling. The `Buffer<N>` abstracts away
almost all unsafe-ness.
The interesting part in this design is the usage of a `ValidatedOffset`
type: this offset is yielded by the `EntryIter`. Since this type knows,
that the returned offset contains a valid `Entry`, it is safe to read.
Therefore the []-indexing is is supported on those special offsets.

There are many new tests added as well.