Only split large enough blocks on allocations
Previously the code always tried to split allocations and only failed to
do so, when the end of the buffer was reached. The critical part was,
that a block thus was split always when it is in the middle, as it is
common for allocations, that will be reused. So, every reused allocation
is split into a USED block and a following FREE block. But what happens
if the block, that this split is not large enough for a USED block and
(even a 0-sized payload) FREE block? Then the splitting clearly should
not happen.

But, the current code splits regardless of that. That overwrites the
following USED block with the information of the "free" block. This
makes the allocation state unusable, since now random parts of heap mem-
ory might be read or written to, even uninitialized memory! This issue
was hidden by a not less important bug: `entry_size - n -HEADER_SIZE`
will subtract with overflow in such cases and happily panic in debug
mode (which is [forbidden] by the current contract of
[`GlobalAlloc`][globalalloc]) and produce a huge number in release mode.
Both renders the allocator unusable as it either violated its safety
contract or simply has an invalid state.

This commit fixes that by simply adding a test, if the block is large
enough for splitting. A proper test case, that fails without the change,
is added as well.

[forbidden]: https://doc.rust-lang.org/1.63.0/core/alloc/trait.GlobalAlloc.html#safety
[globalalloc]: https://doc.rust-lang.org/1.63.0/core/alloc/trait.GlobalAlloc.html