Add a real-world allocation test case
This commit adds a simple but important test case: it uses the allocator
on the machine running the tests. Since this is a hosted environment,
allocations are rather frequent and the heap space is relatively large
(here 4M). This way all the allocations in the `std` and in the test
runner are also directed to `emballoc`, which is great for testing.

This commit also contains a minor adjustment: the README and crate-level
documentation previously mentioned the availability of `HashMap`: this
is not true (most likely since `alloc` still has no random generator re-
quired for the random state of the HashMap) and thus replaced with the
`BTreeMap`.