Test the code snippets of the Readme as doctests
The Readme features the first code snippet a potential user might see of
this crate. Therefore it is highly important, that the code snippet ac-
tually compiles and works. This was not ensured by testing, though!
This commit uses a [trick show be G. Gomez][trick] to include the Readme
as documentation of an `extern`-block, that is only compiled for doc-
tests. This will ensure, that the Readme will continue to provide wor-
king code examples.

[trick]: https://blog.guillaume-gomez.fr/articles/2020-03-07+cfg%28doctest%29+is+stable+and+you+should+use+it