Add a test case for #30
This test case "parses" the test-case binary itself by obtaining the
addresses of the symbols. Then it ensures, that the `.bss`-section comes
directly after the `.data`-section. With this knowledge one can simply
check, that the allocator is placed after the `.bss`-start (so either in
the `.bss`-section itself or in a hypothetical `.noinit`-section, etc.).

This test case currently is failing, as the allocator _is_ placed in the
`.data`-section. See [#30] for details.

[#30]: https://github.com/jfrimmel/emballoc/issues/30