Update and rebuild docker images
This commit updates the Rust versions in the docker image. As the `miri`
Dockerfile overrides the Rust version anyways, this is changed to the
latest docker version at build time. Since the `coverage`-image needs a
custom `PATH`-variable containing the Rust version, it is easiest to
exactly specify the Rust version to use.

Both images now run cargo versions, that support the "sparse registry".

The images are built like this:

    $ docker build -t jfrimmel/coverage:1.1 -f .circleci/coverage.Dockerfile .circleci/
    $ docker tag jfrimmel/coverage:1.1 jfrimmel/coverage:latest
    $ docker push jfrimmel/coverage:1.1
    $ docker push jfrimmel/coverage:latest

    $ docker build -t jfrimmel/miri:nightly-2023-11-09 --build-arg nightly_version=2023-11-09 -f .circleci/miri.Dockerfile .circleci/
    $ docker push jfrimmel/miri:nightly-2023-11-09