Use newly published docker images in CI
The current builds all have a problem: the duration! Half of the jobs
are fast, but the others are hopelessly slow. An example is [build#47],
which has those timings:

    style      4s
    msrv   4m 53s     ┌─── coverage        4m 49s
    doc        3s     ├─── miri-mips64-... 5m 09s
    build      3s ────┼─── miri-x86_64-... 5m 16s
                      └─── test                5s

As you can see, the `coverage`, `msrv` and `miri` jobs are much slower.
Interestingly, exactly those jobs use custom docker images with older
Rust versions (1.62). Those versions crucially don't include the newer
[sparse registry] feature. Investigation shows, that most of those ~5min
is spent on downloading the package index.

This commit uses newer images, which should be vastly faster.

There remains one problem: the `msrv`-job will still be slow, as this
has to use the old compiler version, which does not support the sparse
registry feature yet.

[build#47]: https://app.circleci.com/pipelines/github/jfrimmel/emballoc/47
[sparse registry]: https://blog.rust-lang.org/2023/03/09/Rust-1.68.0.html