Cache the cargo registry
The older Rust 1.57-version is the slowest CI job (by far) and spends
the majority of its time downloading the cargo registry due to the older
git-based protocol. Since this job and the specific Rust version is re-
quired to stay, the build job is speed up by caching the cargo registry.
This decreases the duration of subsequent builds by over 90%.