Reformat assertions
This is necessary to get the line based coverage up. Since both assert-
ions are passing (as expected), the messages are never actually executed
and thus will reported as missing in the coverage report. This is ugly,
but necessary. Note, that this only "improves" the line-based coverage,
other coverage metrics still show the messages to be missed.